+++ modernRX/aes.hpp
@@ -11,9 +11,9 @@
 namespace modernRX::aes {
 	// Performs AES encoding of an input state with given key.
 	// Overwrites input state with encoding result.
-	void encode(std::span<uint32_t, 4> state, const_span<uint32_t, 4> key);
+	void encode(std::span<uint32_t, 4> state, const_span<uint32_t, 4> key) noexcept;
 
 	// Performs AES decoding of an input state with given key.
 	// Overwrites input state with decoding result.
-	void decode(std::span<uint32_t, 4> state, const_span<uint32_t, 4> key);
+	void decode(std::span<uint32_t, 4> state, const_span<uint32_t, 4> key) noexcept;
 }