+++ modernRX/hasher.cpp
@@ -12,12 +12,13 @@ namespace modernRX {
     }
 
     void Hasher::reset(const_span<std::byte> key) {
-        constexpr uint32_t programs_count{ Rx_Cache_Accesses }; // Number of superscalar programs should be equal to number of cache accesses.
+        constexpr uint32_t Programs_Count{ Rx_Cache_Accesses }; // Number of superscalar programs should be equal to number of cache accesses.
 
         if (!this->key.empty() && std::equal(key.begin(), key.end(), this->key.begin())) {
             return;
         }
 
+        this->key.clear();
         this->key.reserve(key.size());
         std::copy(key.begin(), key.end(), std::back_inserter(this->key));
 
@@ -27,7 +28,7 @@ namespace modernRX {
         blake2b::Random blakeRNG{ key, 0 };
         Superscalar superscalar{ blakeRNG };
 
-        std::array<SuperscalarProgram, programs_count> programs;
+        std::array<SuperscalarProgram, Programs_Count> programs;
         for (auto& program : programs) {
             program = superscalar.generate();
         }