+++ modernRX/intrinsics.hpp
@@ -2,19 +2,19 @@
 
 /*
 * Wrapper over compiler intrinsics.
-* Used, but not defined by RandomX algorithm.
+* Used by RandomX algorithm.
 */
 
 #include <intrin.h>
 
 namespace modernRX::intrinsics {
-	inline uint64_t smulh(const int64_t a, const int64_t b) {
+	[[nodiscard]] inline uint64_t smulh(const int64_t a, const int64_t b) noexcept {
 		int64_t hi;
 		_mul128(a, b, &hi);
 		return hi;
 	}
 
-	inline uint64_t umulh(const uint64_t a, const uint64_t b) {
+	[[nodiscard]] inline uint64_t umulh(const uint64_t a, const uint64_t b) noexcept {
 		return __umulh(a, b);
 	}
 }