+++ modernRX/sse.hpp
@@ -3,7 +3,7 @@
 /*
 * Wrapper over some SSE intrinsics required by RandomX algorithm.
 * Implements FloatEnvironment RAII class to set and reset float environment, which is required by RandomX algorithm.
-* Used, but not defined by RandomX algorithm.
+* Used by RandomX algorithm.
 * Code may be a little bit messy and not fully documented as this will be further extended in unknown direction.
 */
 
@@ -22,7 +22,7 @@ namespace modernRX::intrinsics::sse {
         // Swaps current float environment with provided one.
         [[nodiscard]] explicit FloatEnvironment(const uint32_t csr = Rx_Mxcsr_Default) noexcept {
             fegetenv(&fenv);
-            _mm_setcsr(csr); // reset float state
+            _mm_setcsr(csr); // Reset float state.
         }
         ~FloatEnvironment() {
             fesetenv(&fenv);
@@ -41,7 +41,6 @@ namespace modernRX::intrinsics::sse {
         _mm_setcsr(csr | (mode << 13));
     }
 
-
     template<typename T>
     struct xmm_wrapper {
         using type = std::conditional_t<std::is_same_v<T, double>, __m128d, void>;
@@ -53,8 +52,8 @@ namespace modernRX::intrinsics::sse {
     using xmm = xmm_wrapper<T>::type;
 
     template<typename T>
-    [[nodiscard]] xmm<T> vxor(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vxor(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_xor_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -62,8 +61,8 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    [[nodiscard]] xmm<T> vand(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vand(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_and_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -72,8 +71,8 @@ namespace modernRX::intrinsics::sse {
 
 
     template<typename T>
-    [[nodiscard]] xmm<T> vor(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vor(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_or_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -81,17 +80,8 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    void vstore(uintptr_t addr, xmm<T> x) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
-            return _mm_store_pd(addr, x);
-        } else {
-            static_assert(!sizeof(T), "the only supported type for this operation is: float64");
-        }
-    }
-
-    template<typename T>
-    [[nodiscard]] xmm<T> vadd(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vadd(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_add_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -99,8 +89,8 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    [[nodiscard]] xmm<T> vsub(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vsub(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_sub_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -108,8 +98,8 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    [[nodiscard]] xmm<T> vmul(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vmul(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_mul_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -117,8 +107,8 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    [[nodiscard]] xmm<T> vdiv(xmm<T> x, xmm<T> y) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vdiv(const xmm<T> x, const xmm<T> y) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_div_pd(x, y);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -126,8 +116,8 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    [[nodiscard]] xmm<T> vsqrt(xmm<T> x) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vsqrt(const xmm<T> x) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_sqrt_pd(x);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
@@ -135,14 +125,13 @@ namespace modernRX::intrinsics::sse {
     }
 
     template<typename T>
-    [[nodiscard]] xmm<T> vswap(xmm<T> x) noexcept {
-        if constexpr (std::is_same<T, double>::value) {
+    [[nodiscard]] constexpr xmm<T> vswap(const xmm<T> x) noexcept {
+        if constexpr (std::is_same_v<T, double>) {
             return _mm_shuffle_pd(x, x, 1);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: float64");
         }
     }
-
 
     template<typename T1, typename T2>
     concept equal_greater_size = sizeof(T1) >= sizeof(T2);
@@ -150,20 +139,19 @@ namespace modernRX::intrinsics::sse {
     // Converts 2 packed int32 values into two floating point values.
     template<typename To>
         requires std::floating_point<To>&& equal_greater_size<To, int32_t>
-    [[nodiscard]] xmm<To> vcvtpi32(const std::byte* addr) noexcept {
+    [[nodiscard]] constexpr xmm<To> vcvtpi32(const std::byte* addr) noexcept {
         if constexpr (std::is_same_v<To, double>) {
-            __m128i x{ _mm_loadl_epi64(reinterpret_cast<const __m128i*>(addr)) };
+            const __m128i x{ _mm_loadl_epi64(reinterpret_cast<const __m128i*>(addr)) };
             return _mm_cvtepi32_pd(x);
         } else {
             static_assert(!sizeof(To), "the only supported type for this operation is: int32 -> float64");
         }
     }
 
-
     template<typename T>
-    [[nodiscard]] xmm<T> vbcasti64(uint64_t x) noexcept {
+    [[nodiscard]] constexpr xmm<T> vbcasti64(const uint64_t x) noexcept {
         if constexpr (std::is_same_v<T, double>) {
-            __m128i v{ _mm_set1_epi64x(x) };
+            const __m128i v{ _mm_set1_epi64x(x) };
             return _mm_castsi128_pd(v);
         } else {
             static_assert(!sizeof(T), "the only supported type for this operation is: int64 -> float64");